feat: Add test metadata directives and structure tests
Assembler enhancements:
- Implemented .maxcalldepth directive for recursion control
- Implemented .timeout directive for test time limits
- Implemented .skip directive for conditional test execution
- Added skip_newlines() after metadata parsing
- Metadata directives properly filtered during test assembly

Test framework improvements:
- Added structure_unit_tests.cpp for program structure validation
- Enhanced in_assembly_test_executor with metadata support
- Added test validator for structural correctness
- Tests now support max call depth configuration

Configuration updates:
- Added test-related config flags
- Enhanced CPU with call depth tracking

Build system:
- Updated run_tests.sh for new test organization

Result: Comprehensive test metadata support with proper
directive handling and execution control.

Tests: ✅ 114 assembly tests (112 passed, 2 skipped), ✅ 118/118 unit tests

+++ src/test/in_assembly_test_validator.hpp
@@ -82,7 +82,7 @@ public:
             }
         }
         
-        const char* summary_color = (valid_count == results.size()) ? "\033[32m" : "\033[33m";
+        const char* summary_color = (valid_count == static_cast<int>(results.size())) ? "\033[32m" : "\033[33m";
         fmt::print("{}\nTests validated: {} / {}\033[0m\n", 
                   summary_color, valid_count, results.size());
         fmt::print("\033[33mNote: Full test execution not yet implemented - only validation performed\033[0m\n");
